Renamed releaser to multimod
