+++ .gitignore
@@ -9,4 +9,4 @@ Thumbs.db
 coverage.*
 
 semconvgen/semconvgen
-releaser/releaser
+multimod/multimod
