+++ multimod/cmd/tag.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"go.opentelemetry.io/build-tools/releaser/internal/tag"
+	"go.opentelemetry.io/build-tools/multimod/internal/tag"
 )
 
 var (
