+++ multimod/internal/common/doc.go
