+++ multimod/internal/common/test_data/new_module_set_release/versions_invalid_syntax.yaml
