+++ multimod/internal/common/test_data/new_module_versioning/versions_invalid_syntax.yaml
