+++ multimod/internal/common/test_data/verify_git_tags_do_not_already_exist/versions_invalid_syntax.yaml
