+++ multimod/internal/common/tools.go
