+++ multimod/internal/verify/test_data/get_dependencies/versions_invalid_syntax.yaml
